+++ NEWS.md
@@ -12,6 +12,8 @@
 
 - Improved `add_info()`
   - `add_info(.data)` (so no further arguments) now checks to see whether metadata may be added and prompts the user to add it if so
+  - `add_info(.data, optional = TRUE)` extends this to further, optional metadata
+  - This includes prompting the user by layer for multiplex networks
 
 # manynet 2.1.2